Renamed test_hooks because of problems with pytest
